pyproject.toml
[project]
name = "character-details"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "bs4>=0.0.2",
    "ebooklib>=0.19",
    "langchain>=0.3.25",
    "langchain-core>=0.3.60",
    "langchain-openai>=0.3.17",
    "langgraph>=0.4.5",
    "pillow>=11.2.1",
    "pinecone>=7.0.1",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "tqdm>=4.67.1",
]
